Implement basic token parsing
This commit adds basic support for parsing tokens from XML files. I have
not added functionality based on the provided by the DTD File because
it's not clear that preserving this structural data is necessary for the
indexer.

More care should be taken in future to standardise the tokens -- for
instance capitalisation, punctuation, etc.

Signed-off-by: Thomas Passmore <thomas@hateful.dev>