Tweak BM25 parameters, u32 -> u64
I have slightly tweaked the BM25 parameters based on Andrew Trotman &
Matt Crane's "Snip!" paper[1], which offers different k_1 and b
parameters which offer an improvement to the precision of the output.

Secondly, occurances of u32 (unsigned 32-bit integers) in the indexing
code used to refer to "number of X" or "offset from start of file in
bytes" has been replaced with u64. I have done this because varint
encoding means that this only effects the space/time of the programs
when the integers grow large.

Plus, having 32-bit offsets in files imply that each index file is
restricted to ~4GiB in size, which is not great.

[1]: http://www.cs.otago.ac.nz/homepages/andrew/papers/2011-13.pdf

Signed-off-by: Thomas Passmore <thomas@hateful.dev>