Fix search non-determinism, remove OrderedFloat
Rust doesn't implement Ord for floats, therefore they can't be used in
ordered structures like BTrees. To solve this, OrderedFloat was used.
For some reason the ordering/construction of ordered floats in the
`search` function caused a significant amount of non-determinism, where
sometimes queries would not return the same weights.

Instead, weights are collected into a vector and sorted in-place using
PartialOrd. The performance is a little lower, although there were less
results previously.

The OrderedFloat dependency was removed.

Signed-off-by: Thomas Passmore <thomas@hateful.dev>